refactor: bar chart
