+++ apps/dashboard/components/logs/chart/components/logs-chart-error.tsx
