+++ apps/dashboard/components/logs/chart/utils/calculate-timepoints.ts
